Risk and Returns: The Sharpe Ratio
The Sharpe Ratio, developed by Nobel Prize winner William Sharpe some 50 years ago, does precisely this: it compares the return of an investment to that of an alternative and relates the relative return to the risk of the investment, measured by the standard deviation of returns. In this project, I will apply the Sharpe ratio to real financial data using pandas.